+++ sanity/schemaTypes/index.ts
@@ -1,0 +1,5 @@
+import { categoria } from "./categoria";
+import { loja } from "./loja";
+import { produto } from "./produto";
+
+export const schemaTypes = [produto, loja, categoria];